Fixed issue with deleting poll or room that doesn't exist

+++ express_server/api/databaseFunction/pollFunctions.js
@@ -783,38 +783,41 @@ const deletePoll = async (host_id, room_id, poll_id) => {
                         .collection('polls')
                         .doc(poll_id);
 
-        let optionRef = pollRef.collection('Options');
-        let userOptionRef = pollRef.collection('userOptions');
-        let optionSnap = await optionRef.get();
-        let userOptionSnap = await userOptionRef.get();
-        let pollOrder = await getPollOrder(host_id, room_id);
+        let poll = await pollRef.get();
+        if(poll.exists) {
+            let optionRef = pollRef.collection('Options');
+            let userOptionRef = pollRef.collection('userOptions');
+            let optionSnap = await optionRef.get();
+            let userOptionSnap = await userOptionRef.get();
+            let pollOrder = await getPollOrder(host_id, room_id);
 
-        // delete options collection
-        for (let x = 0; x < optionSnap.docs.length; x++) {
-            await optionRef.doc(optionSnap.docs[x].id).delete();
+            // delete options collection
+            for (let x = 0; x < optionSnap.docs.length; x++) {
+                await optionRef.doc(optionSnap.docs[x].id).delete();
+            }
+
+            // delete userOptions collection
+            for (let x = 0; x < userOptionSnap.docs.length; x++) {
+                await userOptionRef.doc(userOptionSnap.docs[x].id).delete();
+            }
+
+            // delete votes document
+            await pollRef.collection('Votes').doc('votes').delete();
+
+            pollOrder.pending = pollOrder.pending.filter(item => item !== poll_id);
+            pollOrder.closed = pollOrder.closed.filter(item => item !== poll_id);
+            pollOrder.open = pollOrder.open.filter(item => item !== poll_id);
+
+            await firestore
+                    .collection(host_id)
+                    .doc(room_id)
+                    .collection('polls')
+                    .doc('order')
+                    .update(pollOrder);
+
+            // delete poll
+            await pollRef.delete();
         }
-
-        // delete userOptions collection
-        for (let x = 0; x < userOptionSnap.docs.length; x++) {
-            await userOptionRef.doc(userOptionSnap.docs[x].id).delete();
-        }
-
-        // delete votes document
-        await pollRef.collection('Votes').doc('votes').delete();
-
-        pollOrder.pending = pollOrder.pending.filter(item => item !== poll_id);
-        pollOrder.closed = pollOrder.closed.filter(item => item !== poll_id);
-        pollOrder.open = pollOrder.open.filter(item => item !== poll_id);
-
-        await firestore
-                .collection(host_id)
-                .doc(room_id)
-                .collection('polls')
-                .doc('order')
-                .update(pollOrder);
-
-        // delete poll
-        await pollRef.delete();
     } catch(error) {
         //console.log(error);
     }
